requirements.txt
aiofiles==24.1.0
aiogram==3.13.1
aiohttp==3.10.9
APScheduler==3.10.4
asyncpg==0.29.0
attrs==24.2.0
magic-filter==1.0.12
psycopg2-binary==2.9.9
pydantic==2.9.2
pydantic_core==2.23.4
python-dotenv==1.0.1
six==1.16.0
SQLAlchemy==2.0.35
